BUGFIX Bot allowed to call itself
Because most people are not going to create a new account when starting out, I’ve removed the *feature* where the bot ignores itself. This was originally put in due to the !repeat demo command which has now been removed anyway. Any command that should ignore bot messages should have a catch within it directly.

Signed-off-by: Matthew Potter <mdpotter@gmail.com>