Updated to provide better control over non-signed in LCTV viewers
Added a new role of “guest” that can now be flagged for messages and joins. Guests are added to a guests array rather than the viewer array in order to allow for contests or other uses for registered viewers to be handled separately.

Best practice is to check the role of the join object to see if the new viewer is signed in or not.

Signed-off-by: Matthew Potter <mdpotter@gmail.com>